More parallel worker tests
