+++ Makefile.toml
@@ -180,7 +180,7 @@ args = [
 ]
 
 [tasks.worker-executor-tests-group1]
-description = "Runs worker executor tests only (group 1/4)"
+description = "Runs worker executor tests only (group 1/5)"
 env = { "WASMTIME_BACKTRACE_DETAILS" = "1", "RUST_LOG" = "info" }
 script = '''
 cargo test --package golem-worker-executor-base --test integration api -- --nocapture
@@ -189,14 +189,14 @@ cargo test --package golem-worker-executor-base --test integration keyvalue -- -
 '''
 
 [tasks.worker-executor-tests-group2]
-description = "Runs worker executor tests only (group 2/4)"
+description = "Runs worker executor tests only (group 2/5)"
 env = { "WASMTIME_BACKTRACE_DETAILS" = "1", "RUST_LOG" = "info" }
 script = '''
 cargo test --package golem-worker-executor-base --test integration guest_languages -- --nocapture
 '''
 
 [tasks.worker-executor-tests-group3]
-description = "Runs worker executor tests only (group 3/4)"
+description = "Runs worker executor tests only (group 3/5)"
 env = { "WASMTIME_BACKTRACE_DETAILS" = "1", "RUST_LOG" = "info" }
 script = '''
 cargo test --package golem-worker-executor-base --test integration transactions -- --nocapture
@@ -204,12 +204,19 @@ cargo test --package golem-worker-executor-base --test integration wasi -- --noc
 '''
 
 [tasks.worker-executor-tests-group4]
-description = "Runs worker executor tests only (group 4/4)"
+description = "Runs worker executor tests only (group 4/5)"
 env = { "WASMTIME_BACKTRACE_DETAILS" = "1", "RUST_LOG" = "info" }
 script = '''
 cargo test --package golem-worker-executor-base --test integration scalability -- --nocapture
 cargo test --package golem-worker-executor-base --test integration hot_update -- --nocapture
 cargo test --package golem-worker-executor-base --test integration rpc -- --nocapture
+'''
+
+[tasks.worker-executor-tests-group5]
+description = "Runs worker executor tests only (group 5/5)"
+env = { "WASMTIME_BACKTRACE_DETAILS" = "1", "RUST_LOG" = "info" }
+script = '''
+cargo test --package golem-worker-executor-base --test integration guest_languages2 -- --nocapture
 '''
 
 [tasks.integration-tests]
